// src/core/enums/log-enum.ts
export enum LogLevel {
  INFO = "info",
  WARN = "warn",
  ERROR = "error"
}

export enum LogOrigin {
  CHECK_CAMPAIGNS_CRON = "CHECK_CAMPAIGNS_CRON",
  CHECK_CURRENT_SEASON_CRON = "CHECK_CURRENT_SEASON_CRON",
  BUILD_NOTIFICATION_TEMPLATE = "BUILD_NOTIFICATION_TEMPLATE",
  CHECK_EMAIL_GALXE = "CHECK_EMAIL_GALXE",
  SAVE_DEPOSIT_VALUES = "SAVE_DEPOSIT_VALUES"
}